+++ CHANGELOG.md
@@ -25,8 +25,9 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 ### Fixed
 
-- `bufio.Scanner: token too long` errors when trying to add filtering-rule lists
-  with lines over 1024 bytes long ([#6003]).
+- `bufio.Scanner: token too long` and other errors when trying to add
+  filtering-rule lists with lines over 1024 bytes long or containing cosmetic
+  rules ([#6003]).
 
 ### Removed